requirements.txt
APScheduler==3.5.3
beautifulsoup4==4.7.1
certifi==2019.3.9
chardet==3.0.4
Django==2.1.1
django-apscheduler==0.2.13
fake-useragent==0.1.11
get==2018.11.19
idna==2.8
itchat==1.3.10
mysqlclient==1.4.2.post1
Pillow==6.0.0
post==2018.11.20
public==2018.11.20
PyJWT==1.7.1
pypng==0.0.19
PyQRCode==1.2.1
PySocks==1.6.8
pytz==2018.9
query-string==2018.11.20
request==2018.11.20
requests==2.21.0
selenium==3.141.0
six==1.12.0
soupsieve==1.9
twilio==6.25.1
tzlocal==1.5.1
urllib3==1.24.1
wikipedia==1.4.0
